修改

+++ ConsoleApplication1.cpp
@@ -1,9 +1,10 @@
 ﻿/*易错程序*/
-#include <stdlib.h>
+#include <stdio.h>
+
 void main1()
 {
 	int num = 10;
-	printf("%f", (float)num);
+	printf("%f", num);
 }
 int main()
 {
